Fix Natspec and overall documentation (#24)
* Remove unnecessary initializations

- removing only from for loops

* Gas optimisation and simplification

* Standardization

- use `getCollateralInAStrategy` instead of `IStrategy(_strategy).checkBalance(_collateral)` directly

* Gas optimisation and simplification

* No need to pre-calculate `_minUSDSAmt` before `mint`

- `mintView` called from YieldReverse will result in the same value as `mintView` called from VaultCore. Hence, `_minUSDSAmt` can just be passed a 0.

* Replace USDS constant variable with Helpers.USDS

* Check for `block.timestamp < lastRebaseTS + gap` instead of `block.timestamp <= lastRebaseTS + gap`

- Gas optimisation

* Use local variable instead of repeated calculation

* Add interface as inheritance in its contract

* Standardize function param name

* Add PR CI

* Fix PR CI

* Fix PR CI

* Fix PR CI

* Fix PR CI

* Unnecessary addition

* fixed a mistake in the comment for buybackPercentage in YieldReserve.sol

* optimized greater than zero checks for unsigned integers. saves gas

* Fix PR CI

* Fix PR CI

* Final fix PR CI

* Fix gitignore

- Remove interfaces folder from gitignore

* Remove `mythril` from `pip-requirements`

- as mythril install throwing error on `npm i`

* improved test coverage for spaBuyback and yieldReserve contracts.

* Improve coverage

- negative tests for USDs and VaultCore

* Fix test

- typo

* Complete coverage for USDs

* Revert parts of "Gas optimisation and simplification"

This reverts parts of commit c29c96d79aa01d15ef72cc79c0d391ebd44c731e.

* added ++i unchecked block in for loop as no possibility of overflow. saves gas

* Update CI

* Update solhint

* Fix Natspec and overall documentation

* Standardize documentation comments

* Update contracts/buyback/SPABuyback.sol

* Apply suggestions from code review

* Fixes

* Format fmt

* Fix package-lock

* docs(NatSpec): Improve natspec for MasterPriceOracle and IOracle

* refactor(contracts/buyback): Improve documentation

* refactor(contracts/vault): Improve CollateralManager documentation

* improved ChainlinkOracle documentation

* docs(Natspec): Improve USDs Token Docs

* refactor(contracts/token): Add missing natspec params

* improved VaultCore documentation

* improved StargateStrategy documentation

* improved documentation for ChainlinkOracle, StargateStrategy, FeeCalculator and Vaultcore

* improved natspec for ChainlinkOracle and FeeCalculator

* docs(Natspec): Improve USDs, Dripper, RebaseManager NatSpec

* refactor(contracts): Make nat-spec docs consistent

* Updated natspec doc

* format fix

* docs(Natspec): Dripper RebaseManager NatSpec

* fix(MasterPriceOracle): added events and errors natspec comments

---------

+++ contracts/buyback/YieldReserve.sol
@@ -10,10 +10,10 @@ import {IOracle} from "../interfaces/IOracle.sol";
 import {Helpers} from "../libraries/Helpers.sol";
 import {IDripper} from "../interfaces/IDripper.sol";
 
-/// @title YieldReserve of USDs protocol
-/// @notice The contract allows users to swap the supported stable coins against yield earned by USDs protocol
-///         It sends USDs to dripper for rebase, and to Buyback Contract for buyback.
+/// @title YieldReserve of USDs Protocol.
 /// @author Sperax Foundation
+/// @notice This contract allows users to swap supported stable-coins for yield earned by the USDs protocol.
+/// It sends USDs to the Dripper contract for rebase and to the Buyback Contract for buyback.
 contract YieldReserve is ReentrancyGuard, Ownable {
     using SafeERC20 for IERC20;
 
@@ -23,15 +23,18 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         uint160 conversionFactor;
     }
 
-    address public vault;
-    address public oracle;
-    address public buyback;
-    address public dripper;
-    // Percentage of USDs to be sent to Buyback 5000 means 50%
+    // Addresses of key contracts
+    address public vault; // Address of the Vault contract
+    address public oracle; // Address of the Oracle contract
+    address public buyback; // Address of the Buyback contract
+    address public dripper; // Address of the Dripper contract
+
+    // Percentage of USDs to be sent to Buyback (e.g., 5000 for 50%)
     uint256 public buybackPercentage;
 
     mapping(address => TokenData) public tokenData;
 
+    // Events
     event Swapped(
         address indexed srcToken,
         address indexed dstToken,
@@ -50,33 +53,34 @@ contract YieldReserve is ReentrancyGuard, Ownable {
     event SrcTokenPermissionUpdated(address indexed token, bool isAllowed);
     event DstTokenPermissionUpdated(address indexed token, bool isAllowed);
 
+    // Custom error messages
     error InvalidSourceToken();
     error InvalidDestinationToken();
     error AlreadyInDesiredState();
     error TokenPriceFeedMissing();
 
-    /// @notice Constructor of the contract
-    /// @param _buyback Address of Buyback contract
-    /// @param _vault Address of Vault
-    /// @param _oracle Address of Oracle
-    /// @param _dripper Address of the dripper contract
+    /// @notice Constructor of the YieldReserve contract.
+    /// @param _buyback Address of the Buyback contract.
+    /// @param _vault Address of the Vault.
+    /// @param _oracle Address of the Oracle.
+    /// @param _dripper Address of the Dripper contract.
     constructor(address _buyback, address _vault, address _oracle, address _dripper) {
         updateBuybackAddress(_buyback);
         updateVaultAddress(_vault);
         updateOracleAddress(_oracle);
         updateDripperAddress(_dripper);
 
-        /// @dev buybackPercentage is initialized to 50%
-        updateBuybackPercentage(uint256(5000));
+        // Initialize buybackPercentage to 50%
+        updateBuybackPercentage(5000);
     }
 
     // OPERATION FUNCTIONS
 
-    /// @notice Swap function to be called by front end
-    /// @param _srcToken Source / Input token
-    /// @param _dstToken Destination / Output token
-    /// @param _amountIn Input token amount
-    /// @param _minAmountOut Minimum output tokens expected
+    /// @notice Swap function to be called by frontend users.
+    /// @param _srcToken Source/Input token.
+    /// @param _dstToken Destination/Output token.
+    /// @param _amountIn Input token amount.
+    /// @param _minAmountOut Minimum output tokens expected.
     function swap(address _srcToken, address _dstToken, uint256 _amountIn, uint256 _minAmountOut) external {
         return swap({
             _srcToken: _srcToken,
@@ -89,9 +93,9 @@ contract YieldReserve is ReentrancyGuard, Ownable {
 
     // ADMIN FUNCTIONS
 
-    /// @notice A function to allow or disallow a `_token` as source token
-    /// @param _token Address of the token
-    /// @param _isAllowed If True, allow it to be used as src token / input token else don't allow
+    /// @notice Allow or disallow a specific `token` for use as a source/input token.
+    /// @param _token Address of the token to be allowed or disallowed.
+    /// @param _isAllowed If set to `true`, the token will be allowed as a source/input token; otherwise, it will be disallowed.
     function toggleSrcTokenPermission(address _token, bool _isAllowed) external onlyOwner {
         TokenData storage data = tokenData[_token];
         if (data.srcAllowed == _isAllowed) revert AlreadyInDesiredState();
@@ -107,9 +111,10 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         emit SrcTokenPermissionUpdated(_token, _isAllowed);
     }
 
-    /// @notice A function to allow or disallow a `_token` as output/destination token.
-    /// @param _token Address of the token
-    /// @param _isAllowed If True, allow it to be used as src token / input token else don't allow
+    /// @notice Allow or disallow a specific `token` for use as a destination/output token.
+    /// @dev Reverts if caller is not owner.
+    /// @param _token Address of the token to be allowed or disallowed.
+    /// @param _isAllowed If set to `true`, the token will be allowed as a destination/output token; otherwise, it will be disallowed.
     function toggleDstTokenPermission(address _token, bool _isAllowed) external onlyOwner {
         TokenData storage data = tokenData[_token];
         if (data.dstAllowed == _isAllowed) revert AlreadyInDesiredState();
@@ -125,21 +130,20 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         emit DstTokenPermissionUpdated(_token, _isAllowed);
     }
 
-    /// @notice Emergency withdrawal function for unexpected situations
-    /// @param _token Address of the asset to be withdrawn
-    /// @param _receiver Address of the receiver of tokens
-    /// @param _amount Amount of tokens to be withdrawn
+    /// @notice Emergency withdrawal function for unexpected situations.
+    /// @param _token Address of the asset to be withdrawn.
+    /// @param _receiver Address of the receiver of tokens.
+    /// @param _amount Amount of tokens to be withdrawn.
     function withdraw(address _token, address _receiver, uint256 _amount) external onlyOwner {
         Helpers._isNonZeroAmt(_amount);
         IERC20(_token).safeTransfer(_receiver, _amount);
         emit Withdrawn(_token, _receiver, _amount);
     }
 
-    /// @notice Set the % of minted USDs sent Buyback
-    /// @param _toBuyback % of USDs sent to Buyback
-    /// @dev The rest of the USDs is sent to VaultCore
-    /// @dev E.g. _toBuyback == 3000 means 30% of the newly
-    ///        minted USDs would be sent to Buyback; the rest 70% to VaultCore
+    /// @notice Set the percentage of newly minted USDs to be sent to the Buyback contract.
+    /// @dev Reverts if caller is not owner.
+    /// @param _toBuyback The percentage of USDs sent to Buyback (e.g., 3000 for 30%).
+    /// @dev The remaining USDs are sent to VaultCore for rebase.
     function updateBuybackPercentage(uint256 _toBuyback) public onlyOwner {
         Helpers._isNonZeroAmt(_toBuyback);
         Helpers._isLTEMaxPercentage(_toBuyback);
@@ -147,44 +151,49 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         emit BuybackPercentageUpdated(_toBuyback);
     }
 
-    /// @notice Update Buyback contract's address
-    /// @param _newBuyBack New address of Buyback contract
+    /// @notice Update the address of the Buyback contract.
+    /// @dev Reverts if caller is not owner.
+    /// @param _newBuyBack New address of the Buyback contract.
     function updateBuybackAddress(address _newBuyBack) public onlyOwner {
         Helpers._isNonZeroAddr(_newBuyBack);
         buyback = _newBuyBack;
         emit BuybackAddressUpdated(_newBuyBack);
     }
 
-    /// @notice Update Oracle's address
-    /// @param _newOracle New address of Oracle
+    /// @notice Update the address of the Oracle contract.
+    /// @dev Reverts if caller is not owner.
+    /// @param _newOracle New address of the Oracle contract.
     function updateOracleAddress(address _newOracle) public onlyOwner {
         Helpers._isNonZeroAddr(_newOracle);
         oracle = _newOracle;
         emit OracleUpdated(_newOracle);
     }
 
-    /// @notice Update Dripper's address
-    /// @param _newDripper New address of Dripper
+    /// @notice Update the address of the Dripper contract.
+    /// @dev Reverts if caller is not owner.
+    /// @param _newDripper New address of the Dripper contract.
     function updateDripperAddress(address _newDripper) public onlyOwner {
         Helpers._isNonZeroAddr(_newDripper);
         dripper = _newDripper;
         emit DripperAddressUpdated(_newDripper);
     }
 
-    /// @notice Update VaultCore's address
-    /// @param _newVault New address of VaultCore
+    /// @notice Update the address of the VaultCore contract.
+    /// @dev Reverts if caller is not owner.
+    /// @param _newVault New address of the VaultCore contract.
     function updateVaultAddress(address _newVault) public onlyOwner {
         Helpers._isNonZeroAddr(_newVault);
         vault = _newVault;
         emit VaultAddressUpdated(_newVault);
     }
 
-    /// @notice Swap allowed src token to allowed dst token.
-    /// @param _srcToken Source / Input token
-    /// @param _dstToken Destination / Output token
-    /// @param _amountIn Input token amount
-    /// @param _minAmountOut Minimum output tokens expected
-    /// @param _receiver Receiver of the tokens
+    /// @notice Swap allowed source token for allowed destination token.
+    /// @dev Reverts if caller is not owner.
+    /// @param _srcToken Source/Input token.
+    /// @param _dstToken Destination/Output token.
+    /// @param _amountIn Input token amount.
+    /// @param _minAmountOut Minimum output tokens expected.
+    /// @param _receiver Receiver of the tokens.
     function swap(address _srcToken, address _dstToken, uint256 _amountIn, uint256 _minAmountOut, address _receiver)
         public
         nonReentrant
@@ -199,8 +208,7 @@ contract YieldReserve is ReentrancyGuard, Ownable {
             // Mint USDs
             IERC20(_srcToken).forceApprove(vault, _amountIn);
             IVault(vault).mint(_srcToken, _amountIn, 0, block.timestamp);
-            // No need to do slippage check as it is our contract
-            // and the vault does that.
+            // No need to do a slippage check as it is our contract, and the vault does that.
         }
         IERC20(_dstToken).safeTransfer(_receiver, amountToSend);
         _sendUSDs();
@@ -213,10 +221,12 @@ contract YieldReserve is ReentrancyGuard, Ownable {
         });
     }
 
-    /// @notice A `view` function to get estimated output
-    /// @param _srcToken Input token address
-    /// @param _dstToken Output token address
-    /// @param _amountIn Input amount of _srcToken
+    /// @notice Get an estimate of the output token amount for a given input token amount.
+    /// @dev Reverts if caller is not owner.
+    /// @param _srcToken Input token address.
+    /// @param _dstToken Output token address.
+    /// @param _amountIn Input amount of _srcToken.
+    /// @return Estimated output token amount.
     function getTokenBForTokenA(address _srcToken, address _dstToken, uint256 _amountIn)
         public
         view
@@ -239,15 +249,15 @@ contract YieldReserve is ReentrancyGuard, Ownable {
 
     // UTILITY FUNCTIONS
 
-    /// @notice Sends USDs to buyback as per buybackPercentage
-    ///         and rest to VaultCore for rebase
+    /// @notice Distributes USDs to the Buyback and Dripper contracts based on buybackPercentage.
+    /// @dev Sends a portion of the USDs balance to the Buyback contract and the remaining to the Dripper contract for rebase.
     function _sendUSDs() private {
         uint256 balance = IERC20(Helpers.USDS).balanceOf(address(this));
 
-        // Calculating the amount to send to Buyback based on buybackPercentage
+        // Calculate the amount to send to Buyback based on buybackPercentage
         uint256 toBuyback = (balance * buybackPercentage) / Helpers.MAX_PERCENTAGE;
 
-        // Remaining balance will be sent to Dripper for rebase
+        // The remaining balance is sent to the Dripper for rebase
         uint256 toDripper = balance - toBuyback;
 
         emit USDsSent(toBuyback, toDripper);
